requirements.txt
streamlit
pdf2docx
python-docx
